Message notifier refactor.
